requirements.txt
fastapi
uvicorn
requests
python-dotenv
firebase-admin
google-cloud-firestore
pydantic
SQLAlchemy
pyodbc
sql